requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
google-cloud-monitoring>=2.15.0
google-cloud-bigquery>=3.10.0
google-generativeai>=0.3.0
gunicorn>=21.2.0
